refactor: rename grpo_loss to prime_rl_loss and consolidate loss interface
- Rename grpo_loss to prime_rl_loss (uses LossConfig directly)
- Move LossInputs, LossOutputs, LossFn from loss_interface.py into loss.py
- Remove loss_interface.py
